+++ ignite/cmd/generate_go.go
@@ -1,27 +1,36 @@
 package ignitecmd
 
 import (
-	"fmt"
-
 	"github.com/spf13/cobra"
 
-	"github.com/ignite/cli/ignite/pkg/cliui/clispinner"
+	"github.com/ignite/cli/ignite/pkg/cliui"
 	"github.com/ignite/cli/ignite/services/chain"
 )
 
 func NewGenerateGo() *cobra.Command {
-	return &cobra.Command{
-		Use:   "proto-go",
-		Short: "Generate proto based Go code needed for the app's source code",
-		RunE:  generateGoHandler,
+	c := &cobra.Command{
+		Use:     "proto-go",
+		Short:   "Generate proto based Go code needed for the app's source code",
+		PreRunE: gitChangesConfirmPreRunHandler,
+		RunE:    generateGoHandler,
 	}
+
+	c.Flags().AddFlagSet(flagSetYes())
+
+	return c
 }
 
 func generateGoHandler(cmd *cobra.Command, args []string) error {
-	s := clispinner.New().SetText("Generating...")
-	defer s.Stop()
+	session := cliui.New(cliui.StartSpinner())
+	defer session.End()
 
-	c, err := newChainWithHomeFlags(cmd)
+	session.StartSpinner("Generating...")
+
+	c, err := newChainWithHomeFlags(
+		cmd,
+		chain.WithOutputer(session),
+		chain.CollectEvents(session.EventBus()),
+	)
 	if err != nil {
 		return err
 	}
@@ -35,8 +44,5 @@ func generateGoHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	s.Stop()
-	fmt.Println("⛏️  Generated go code.")
-
-	return nil
+	return session.Println("⛏️  Generated go code.")
 }