+++ ignite/cmd/network_chain_prepare.go
@@ -41,8 +41,8 @@ func NewNetworkChainPrepare() *cobra.Command {
 }
 
 func networkChainPrepareHandler(cmd *cobra.Command, args []string) error {
-	session := cliui.New()
-	defer session.Cleanup()
+	session := cliui.New(cliui.StartSpinner())
+	defer session.End()
 
 	force, _ := cmd.Flags().GetBool(flagForce)
 
@@ -109,7 +109,6 @@ func networkChainPrepareHandler(cmd *cobra.Command, args []string) error {
 	}
 	binaryDir := filepath.Dir(filepath.Join(goenv.Bin(), binaryName))
 
-	session.StopSpinner()
 	session.Printf("%s Chain is prepared for launch\n", icons.OK)
 	session.Println("\nYou can start your node by running the following command:")
 	commandStr := fmt.Sprintf("%s start --home %s", binaryName, chainHome)