+++ ignite/services/chain/generate.go
@@ -2,7 +2,6 @@ package chain
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"path/filepath"
 
@@ -134,7 +133,7 @@ func (c *Chain) Generate(
 		return err
 	}
 
-	fmt.Fprintln(c.stdLog().out, "🛠️  Building proto...")
+	c.ev.Send("🛠  Building proto...")
 
 	options := []cosmosgen.Option{
 		cosmosgen.IncludeDirs(conf.Build.Proto.ThirdPartyPaths),